FTP에서 커맨드 라인 인자 입력 완성
